booksGen/pyproject.toml
[tool.poetry]
name = "booksgen"
version = "0.1.0"
description = ""
authors = ["Josué Luiz Barbosa e Silva <104951932+josuelb@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.2"}
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
pydantic-settings = "^2.5.2"
pyjwt = "^2.9.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.9"
tzdata = "^2024.1"
redis = "^5.1.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.3"
pytest = "^8.3.2"
pymysql = "^1.1.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
